+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added 
+- Building direct, structural economic loss analysis.  [INCORE1-442](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-442).
+
 ## [0.7.0] - 2020-07-31
 
 ### Added